+++ frontend/src/components/log-in/log-in.module.css
@@ -18,6 +18,7 @@
 
 .text {
   margin: 0;
+  text-decoration: none;
   font-family: 'IBM Plex Sans', sans-serif;
   font-weight: 500;
   font-size: 20px;
