+++ frontend/src/components/log-in/log-in.tsx
@@ -1,16 +1,15 @@
-import { useState, useContext } from "react";
+import { Link } from "react-router-dom";
 import styles from "./log-in.module.css";
 
-export const LogIn = (props: {auth: () => any}) => {
-  const { auth } = props;
-  
-
+export const LogIn = () => {
   return (
     <div className={styles.section}>
       <div className={styles.container}>
         <h1 className={styles.header}>С кем я учусь?</h1>
-        <button className={styles.button} onClick={auth}>
-          <p className={styles.text}>Войти с Яндекс ID</p>
+        <button className={styles.button}>
+          <Link className={styles.text} to='/switch-profile'>
+            Войти с Яндекс ID
+          </Link>
         </button>
       </div>
     </div>
