+++ frontend/src/index.tsx
@@ -5,7 +5,7 @@ import "./fonts/fonts.css";
 import App from "./components/app/app";
 import { BrowserRouter } from "react-router-dom";
 import reportWebVitals from "./reportWebVitals";
-import { AppProvider } from "./AppContext";
+import { AppProvider } from "./utils/AppContext";
 
 ReactDOM.render(
   <React.StrictMode>
