+++ frontend/src/utils/AppContext.tsx
@@ -1,6 +1,6 @@
 import { createContext, Dispatch, useReducer } from 'react';
-import { Action, initialState, reducer, User } from './reducer';
-import { State } from './reducer'
+import { Action, initialState, reducer, User } from '../reducer';
+import { State } from '../reducer'
 
 const AppContext = createContext<{
   state: State;
